requirements.txt
attrs==19.1.0
backcall==0.1.0
bokeh==1.3.4
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
coverage==4.5.4
decorator==4.4.0
Django==2.2.5
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
idna==2.8
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.17.2
packaging==19.1
pandas==0.25.1
pandas-highcharts==0.5.2
parso==0.5.1
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
Pygments==2.4.2
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
tornado==6.0.3
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
